BUG fix PR json validation for no PR made
