+++ tests/model/test_validate.py
@@ -218,4 +218,4 @@ def test_model_invalid(model: PerPackageModel, invalid_feedstock: str):
 
 
 def test_validate_pr_json(pr_json: str):
-    PullRequestData.model_validate_json(pr_json)
+    PullRequestData.validate_json(pr_json)
